[Refactor] Firebase DataSource 역할 분리
